Add a VBScript export tool for the salary CSV pipeline, fix spok's real type
VBS twin of export-salary-views-to-csv.ps1, for running via Windows Task
Scheduler without PowerShell execution-policy setup. Uses ADODB
(Connection/Recordset/Stream) to hit the same two Access views, writing
identical CSV output (UTF-8 no BOM, quoted fields, bare \N for NULL).

Testing it surfaced a real bug the PowerShell version had been masking:
wv_datein_group_week.spok is genuinely a Double in the source data (e.g.
24463.6596679688), not an integer count as the original schema assumed.
MySQL silently truncated it with just a warning when fed from the
PowerShell CSV (period decimal separator, still parseable as a number
prefix) -- but VBScript's CStr() on a Double uses the machine's regional
settings, and a Russian-locale Windows renders it as "24463,6596679688"
(comma), which MySQL can't parse as a number at all, turning silent data
loss into a hard import failure. Fixed the column to DOUBLE
(migrations/007) and made the VBS format numeric fields with an explicit
'.' regardless of locale.

The .vbs file itself is saved as UTF-16LE with a BOM, not UTF-8 --
cscript.exe misreads non-ASCII bytes (e.g. in a Cyrillic path) in a
plain-UTF-8 .vbs file.

+++ migrations/salary_schema.sql
@@ -69,6 +69,6 @@ CREATE TABLE wv_datein_group_week (
     pweekno VARCHAR(20) NOT NULL,
     phone VARCHAR(32) NULL,
     id INT NOT NULL,
-    spok INT NULL,
+    spok DOUBLE NULL,
     KEY idx_week_phone_id (pweekno, phone, id)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
